backend/requirements.txt
sentence-transformers==2.2.2
langchain
python-dotenv
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_experimental
ctransformers==0.2.5
fastapi
uvicorn
-e .